@@ -64,8 +64,8 @@
     "rollup-plugin-filesize": "9.0.2",
     "rollup-plugin-node-resolve": "5.2.0",
     "rollup-plugin-terser": "7.0.2",
-    "vue": "2.6.11",
-    "vue-class-component": "7.2.5",
+    "vue": "3.0.0",
+    "vue-class-component": "^8.0.0-0",
     "vue-template-compiler": "2.6.11",
     "vuex": "3.1.1"
   },
